catalog: extract catalog_data.py — data literals + dataclasses (#38 step 2)
Step 2 of the catalog.py split: move the 92 static-data symbols (Provider/Model/
ModelEndpoint dataclasses, PROVIDERS, privacy-tier + orchestration constants, all
*_MODEL_ID + *_MODEL_ORDER, embedding specs, allowlists) into catalog_data.py.
Verified a clean leaf — depends only on pricing (PriceTier for Model.price_tiers),
NEVER on the live MODELS/MODEL_ENDPOINTS registries (which stay in catalog.py,
built at import from this data + the ingested snapshot). catalog.py re-imports
for back-compat, so all importers are unaffected.

catalog.py: 3374 -> 2119 lines. Facade verified (Model/PROVIDERS/META_MODEL_IDS
re-export; MODELS/ENDPOINTS still build 270/701). Full suite 1175 green, mypy clean.
